+++ packages/@valapi/lib/src/client/ValRegion.ts
@@ -4,13 +4,13 @@ export class ValRegion {
     /**
      * Region ID
      */
-    public readonly id: Region.Identify;
+    public readonly id: Region.ID;
 
     /**
      *
-     * @param {Region.Identify} region Region (default: na)
+     * @param {Region.ID} region Region (default: na)
      */
-    public constructor(region: Region.Identify = Region.Default.North_America) {
+    public constructor(region: Region.ID = Region.Default.North_America) {
         this.id = "na";
         switch (region) {
             case "na": {