+++ packages/@valapi/lib/src/resources/CrosshairColor.ts
@@ -1,112 +1,76 @@
-export default class Data {
-    public static readonly from = {
-        "0": "White",
-        "1": "Green",
-        "2": "Yellow_Green",
-        "3": "Green_Yellow",
-        "4": "Yellow",
-        "5": "Cyan",
-        "6": "Pink",
-        "7": "Red",
-        "8": "Custom"
-    };
+export const Default = <const>{
+    White: "0",
+    Green: "1",
+    Yellow_Green: "2",
+    Green_Yellow: "3",
+    Yellow: "4",
+    Cyan: "5",
+    Pink: "6",
+    Red: "7",
+    Custom: "8"
+};
 
-    public static readonly fromHex = {
-        "0": "FFFFFF",
-        "1": "00FF00",
-        "2": "7FFF00",
-        "3": "DFFF00",
-        "4": "FFFF00",
-        "5": "00FFFF",
-        "6": "FF00FF",
-        "7": "FF0000"
-    };
-
-    public static readonly to = {
-        White: "0",
-        Green: "1",
-        Yellow_Green: "2",
-        Green_Yellow: "3",
-        Yellow: "4",
-        Cyan: "5",
-        Pink: "6",
-        Red: "7",
-        Custom: "8"
-    };
-
-    public static readonly toHex = {
-        FFFFFF: "0",
-        "00FF00": "1",
-        "7FFF00": "2",
-        DFFF00: "3",
-        FFFF00: "4",
-        "00FFFF": "5",
-        FF00FF: "6",
-        FF0000: "7"
-    };
-}
-
-export enum Default {
-    White = "0",
-    Green = "1",
-    Yellow_Green = "2",
-    Green_Yellow = "3",
-    Yellow = "4",
-    Cyan = "5",
-    Pink = "6",
-    Red = "7",
-    Custom = "8"
-}
-
-export enum DefaultColor {
-    White = "FFFFFF",
-    Green = "00FF00",
-    Yellow_Green = "7FFF00",
-    Green_Yellow = "DFFF00",
-    Yellow = "FFFF00",
-    Cyan = "00FFFF",
-    Pink = "FF00FF",
-    Red = "FF0000"
-}
-
-export type Identify = keyof typeof Data.from;
-export type Color = keyof typeof Data.to;
-
-export type IdentifyHex = keyof typeof Data.fromHex;
-export type ColorHex = keyof typeof Data.toHex;
+export type Color = keyof typeof Default;
+export type ID = (typeof Default)[Color];
 
 /**
- * Change from {@link Color} to {@link Identify String}
+ * Change from {@link Color} to {@link ID}
  * @param {Color} x Color
- * @returns {Identify} String
+ * @returns {ID} ID
  */
-export function fromColor(x: Color): Identify {
-    return <Identify>Data.to[x];
+export function fromColor(x: Color): ID | undefined {
+    return <ID>Default[x];
 }
 
 /**
- * **(Hex)** Change from {@link ColorHex Color} to {@link IdentifyHex String}
- * @param {ColorHex} x Color
- * @returns {IdentifyHex} String
- */
-export function fromColorHex(x: ColorHex): IdentifyHex {
-    return <IdentifyHex>Data.toHex[x];
-}
-
-/**
- * Change from {@link Identify String} to {@link Color}
- * @param {Identify} x String
+ * Change from {@link ID} to {@link Color}
+ * @param {ID} x ID
  * @returns {Color} Color
  */
-export function fromString(x: Identify): Color {
-    return <Color>Data.from[x];
+export function fromID(x: ID): Color | undefined {
+    for (const data of Object.entries(Default)) {
+        if (data[1] == x) {
+            return <Color>data[0];
+        }
+    }
+
+    return undefined;
+}
+
+export const DefaultHex = <const>{
+    White: "FFFFFF",
+    Green: "00FF00",
+    Yellow_Green: "7FFF00",
+    Green_Yellow: "DFFF00",
+    Yellow: "FFFF00",
+    Cyan: "00FFFF",
+    Pink: "FF00FF",
+    Red: "FF0000"
+};
+
+type ColorHex = keyof typeof DefaultHex;
+export type IdHex = (typeof DefaultHex)[ColorHex];
+
+/**
+ * Change from {@link Color} to {@link IdHex}
+ * @param {Color} x Color
+ * @returns {IdHex} IdHex
+ */
+export function fromColorHex(x: Color): IdHex | undefined {
+    return <IdHex>DefaultHex[<ColorHex>x];
 }
 
 /**
- * **(Hex)** Change from {@link IdentifyHex String} to {@link ColorHex Color}
- * @param {IdentifyHex} x String
- * @returns {ColorHex} Color
+ * Change from {@link IdHex} to {@link Color}
+ * @param {IdHex} x IdHex
+ * @returns {Color} Color
  */
-export function fromStringHex(x: IdentifyHex): ColorHex {
-    return <ColorHex>Data.fromHex[x];
+export function fromIdHex(x: IdHex): Color | undefined {
+    for (const data of Object.entries(DefaultHex)) {
+        if (data[1] == x) {
+            return <Color>data[0];
+        }
+    }
+
+    return undefined;
 }