+++ packages/@valapi/lib/src/resources/Region.ts
@@ -1,52 +1,36 @@
-export default class Data {
-    public static readonly from = {
-        na: "North_America",
-        latam: "Latin_America",
-        br: "Brazil",
-        pbe: "Public_Beta_Environment",
-        eu: "Europe",
-        kr: "Korea",
-        ap: "Asia_Pacific"
-    };
+export const Default = <const>{
+    North_America: "na",
+    Latin_America: "latam",
+    Brazil: "br",
+    Public_Beta_Environment: "pbe",
+    Europe: "eu",
+    Korea: "kr",
+    Asia_Pacific: "ap"
+};
 
-    public static readonly to = {
-        North_America: "na",
-        Latin_America: "latam",
-        Brazil: "br",
-        Public_Beta_Environment: "pbe",
-        Europe: "eu",
-        Korea: "kr",
-        Asia_Pacific: "ap"
-    };
-}
-
-export enum Default {
-    North_America = "na",
-    Latin_America = "latam",
-    Brazil = "br",
-    Public_Beta_Environment = "pbe",
-    Europe = "eu",
-    Korea = "kr",
-    Asia_Pacific = "ap"
-}
-
-export type Identify = keyof typeof Data.from;
-export type Name = keyof typeof Data.to;
+export type Name = keyof typeof Default;
+export type ID = (typeof Default)[Name];
 
 /**
- * Change from {@link Name} to {@link Identify String}
+ * Change from {@link Name} to {@link ID}
  * @param {Name} x Name
- * @returns {Identify} String
+ * @returns {ID} ID
  */
-export function fromName(x: Name): Identify {
-    return <Identify>Data.to[x];
+export function fromName(x: Name): ID | undefined {
+    return <ID>Default[x];
 }
 
 /**
- * Change from {@link Identify String} to {@link Name}
- * @param {Identify} x String
+ * Change from {@link ID} to {@link Name}
+ * @param {ID} x ID
  * @returns {Name} Name
  */
-export function fromString(x: Identify): Name {
-    return <Name>Data.from[x];
+export function fromID(x: ID): Name | undefined {
+    for (const data of Object.entries(Default)) {
+        if (data[1] == x) {
+            return <Name>data[0];
+        }
+    }
+
+    return undefined;
 }